feat:Optionfour구조변경
